refactor(fuse/table_functions): rename to history_table
