+++ query/src/storages/fuse/mod.rs
@@ -15,14 +15,14 @@
 pub mod cache;
 mod constants;
 mod fuse_table;
+mod history_table;
 pub mod io;
 pub mod meta;
 pub mod operations;
 pub mod pruning;
 pub mod statistics;
-mod table_functions;
 
 pub use constants::*;
 pub use fuse_table::FuseTable;
-pub use table_functions::FuseHistoryTable;
-pub use table_functions::FUSE_FUNC_HIST;
+pub use history_table::FuseHistoryTable;
+pub use history_table::FUSE_FUNC_HIST;
