+++ query/tests/it/storages/fuse/history_table/fuse_history_table.rs
