+++ query/tests/it/storages/fuse/mod.rs
@@ -12,10 +12,10 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
+mod history_table;
 mod io;
 mod operations;
 mod pruning;
 mod statistics;
 mod table;
-mod table_functions;
 mod table_test_fixture;
